pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "trading"
version = "0.2.0"
description = "Trading bot system with Rust core and Python strategies"
readme = "README.md"
requires-python = ">3.11"
dependencies = [
    "maturin>=1.10.2",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "ruff>=0.8.0",
    "yfinance>=0.2.0",
    "gymnasium>=0.29.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
rl = [
    "stable-baselines3>=2.0.0",
    "torch>=2.0.0",
]

[project.scripts]
trading = "trading.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.maturin]
# Rust code is in the rust/ directory
manifest-path = "rust/Cargo.toml"
module-name = "trading._core"
# Private project - no publishing needed
manylinux = "2014"
